+++ src/main/java/rocks/zipcode/atm/bank/PremiumAccount.java
@@ -5,7 +5,7 @@ package rocks.zipcode.atm.bank;
  */
 public class PremiumAccount extends Account {
 
-        private static final int OVERDRAFT_LIMIT = 100;
+        private static final float OVERDRAFT_LIMIT = 100;
 
         public PremiumAccount(AccountData accountData) {
             super(accountData);
